services/email_service/requirements.txt
Django>=4.2
psycopg2-binary>=2.9
google-auth
djangorestframework-simplejwt
drf-spectacular
drf-yasg
requests>=2.31
pika==1.3.2
django-cors-headers==4.3.1